# docs/jobScheduler/jobScheduler_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for the Job Scheduler service.",
        "title": "Job Scheduler",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/job-scheduler",
    "paths": {
        "/messenger/is-working": {
            "get": {
                "description": "Checks whether the Messenger Job is currently running.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MessengerJob"
                ],
                "summary": "Check Messenger Job Status",
                "responses": {
                    "200": {
                        "description": "Current job status",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
        "/messenger/start": {
            "get": {
                "description": "Initiates the Messenger Job and returns the operation status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MessengerJob"
                ],
                "summary": "Start Messenger Job",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        },
        "/messenger/stop": {
            "get": {
                "description": "Terminates the Messenger Job and returns the operation status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MessengerJob"
                ],
                "summary": "Stop Messenger Job",
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/pkg.BaseResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "pkg.BaseResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "internalCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}